voiceguard-ai/nlp/scrapers/requirements.txt
python-dotenv>=0.21.0
feedparser>=6.0.8
apscheduler>=3.9.1
requests>=2.28.0
beautifulsoup4>=4.12.2
praw>=7.0.0
spacy>=3.0.0
transformers>=4.0.0
